requirements.txt
# PDF extraction
pdfplumber==0.11.4

# Free local embeddings (no API cost)
sentence-transformers==3.3.1

# Vector database
chromadb==0.5.23

# OpenAI API
openai==1.57.0

# Web UI
streamlit==1.41.1

# Loads .env file automatically
python-dotenv==1.0.1

# Progress bars
tqdm==4.67.1
